rpm, not count now

+++ src/main/kotlin/EventsStatistic.kt
@@ -1,5 +1,5 @@
 interface EventsStatistic {
     fun incEvent(name: String)
-    fun getEventsStatisticByName(name: String): Int
-    fun getAllEventStatistic(): Map<String, Int>
+    fun getEventsStatisticByName(name: String): Double
+    fun getAllEventStatistic(): Map<String, Double>
 }
